xmlstream.py: move SSL socket creation code to _create_secure_socket()
Currently in xmlstream.py, there are two pieces of the same code for
creating a SSL socket, one for Legacy SSL, another for StartTLS. A
separate function, XMLStream._create_secure_socket() is introduced
to reuse the same code, which returns a wrapped ssl_socket.

This change makes it easier to further improve SSL/TLS code.

Signed-off-by: Tom Li <biergaizi@member.fsf.org>