Fixes compatibility issues with pyasn1 >= 0.4.1
This is a multipart Problem.

Part 1. Is the year of the certificates. You have provided a solution for this. I have altered this solution because of the second part.

Part 2. There is a traceback the occurs when the TLS certificate is checked. 

    pyasn1.error.PyAsn1Error: <TagSet object at 0x7f53f5fb74e0 tags 0:32:16> not in asn1Spec: <OctetString schema object at 0x7f53f5f51d68 tagSet <TagSet object at 0x7f540c3cef28 tags 0:0:4> encoding iso-8859-1>

This traceback has been reported to the maker(s) of pyasn1.

https://github.com/etingof/pyasn1/issues/112

This issue was closed with a half explained solution. I have instituted this solution into the sleekxmpp library.

I added version checking of the pyasn1 module because of this the checking of the date instances is not needed (Or at least I do not think it is) If the pyasn1 version is >= 0.4.1 then a flag is set to use the alternate means of grabbing the needed certificate information.